Move ListDialogFragments class in fragments package
